feature: add `atomic_file_swap(create_dirs)` parameter

+++ src/kraken/core/utils.py
@@ -4,6 +4,7 @@ import contextlib
 import enum
 import importlib
 import os
+import shutil
 import tempfile
 from pathlib import Path
 from typing import IO, AnyStr, BinaryIO, ContextManager, Iterable, Iterator, TextIO, TypeVar, overload
@@ -33,6 +34,7 @@ def atomic_file_swap(
     path: str | Path,
     mode: Literal["w"],
     always_revert: bool = ...,
+    create_dirs: bool = ...,
 ) -> ContextManager[TextIO]:
     ...
 
@@ -42,6 +44,7 @@ def atomic_file_swap(
     path: str | Path,
     mode: Literal["wb"],
     always_revert: bool = ...,
+    create_dirs: bool = ...,
 ) -> ContextManager[BinaryIO]:
     ...
 
@@ -51,6 +54,7 @@ def atomic_file_swap(
     path: str | Path,
     mode: Literal["w", "wb"],
     always_revert: bool = False,
+    create_dirs: bool = False,
 ) -> Iterator[IO[AnyStr]]:
     """Performs an atomic write to a file while temporarily moving the original file to a different random location.
 
@@ -58,6 +62,8 @@ def atomic_file_swap(
         path: The path to replace.
         mode: The open mode for the file (text or binary).
         always_revert: If enabled, swap the old file back into place even if the with context has no errors.
+        create_dirs: If the file does not exist, and neither do its parent directories, create the directories.
+            The directory will be removed if the operation is reverted.
     """
 
     path = Path(path)
@@ -74,19 +80,26 @@ def atomic_file_swap(
             )
             old.close()
             os.rename(path, old.name)
+        else:
+            old = None
 
-            def _revert() -> None:
-                assert isinstance(path, Path)
-                if path.is_file():
-                    path.unlink()
+        def _revert() -> None:
+            assert isinstance(path, Path)
+            if path.is_file():
+                path.unlink()
+            if old is not None:
                 os.rename(old.name, path)
 
-        else:
+        if not path.parent.is_dir() and create_dirs:
+            path.parent.mkdir(exist_ok=True)
+            _old_revert = _revert
 
             def _revert() -> None:
                 assert isinstance(path, Path)
-                if path.is_file():
-                    path.unlink()
+                try:
+                    shutil.rmtree(path.parent)
+                finally:
+                    _old_revert()
 
         try:
             with path.open(mode) as new:
@@ -98,7 +111,8 @@ def atomic_file_swap(
             if always_revert:
                 _revert()
             else:
-                os.remove(old.name)
+                if old is not None:
+                    os.remove(old.name)
 
 
 @overload
